refactor: Extract VoiceInput from VDF
