+++ app/src/main/java/com/algolia/instantsearch/voice/demo/MainActivity.kt
@@ -5,9 +5,10 @@ import android.support.v7.app.AppCompatActivity
 import com.algolia.instantsearch.voice.Voice
 import com.algolia.instantsearch.voice.ui.PermissionDialogFragment
 import com.algolia.instantsearch.voice.ui.VoiceDialogFragment
+import com.algolia.instantsearch.voice.VoiceInput
 import kotlinx.android.synthetic.main.activity_main.*
 
-class MainActivity : AppCompatActivity(), VoiceDialogFragment.VoiceResultsListener {
+class MainActivity : AppCompatActivity(), VoiceInput.VoiceResultsListener {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
@@ -17,8 +18,9 @@ class MainActivity : AppCompatActivity(), VoiceDialogFragment.VoiceResultsListen
                 PermissionDialogFragment().show(supportFragmentManager, "perm")
             } else {
                 val voiceFragment = VoiceDialogFragment() //FIXME: Handle orientation changes, storing state properly
-                voiceFragment.language = "en-US"
-                voiceFragment.maxResults = 2
+                voiceFragment.setSuggestions("Something", "Something else")
+                voiceFragment.input.language = "en-US"
+                voiceFragment.input.maxResults = 2
                 voiceFragment.show(supportFragmentManager, "voice")
             }
         }
